Fork package as cherrytree-docmodel for side-by-side install
Rename binary, desktop file, icons, man page, metainfo, mime files,
and D-Bus app ID so cherrytree-docmodel can be installed alongside
the upstream cherrytree package without file conflicts.

Use new file extensions (.ctbx, .ctdx, .ctxx, .ctzx) to prevent
standard cherrytree from accidentally opening and corrupting rich
table data via autosave. Legacy extensions still accepted for opening.

Fix fmt compile error with GdkModifierType on Ubuntu 25.04 and add
toolbar migration to inject undo/redo buttons into saved configs.

+++ tests/tests_command_gui_simulation.cpp
@@ -971,7 +971,7 @@ void TestGuiSimulationApp::_test_gtk_accelerator_bindings(CtMainWin* /*pWin*/)
 
         if (found) {
             spdlog::info("  ✓ Found accelerator: {} -> key={}, mods={}",
-                         test.description, gdk_keyval_name(key.accel_key), key.accel_mods);
+                         test.description, gdk_keyval_name(key.accel_key), static_cast<int>(key.accel_mods));
             foundCount++;
 
             // Note: We don't strictly verify the exact key/mod combination because
